# tests/api/openapi_v1_snapshot.json
{
  "components": {
    "schemas": {
...
      "BuildSnapshotRequest": {
        "description": "Request to build a context snapshot from a declarative build spec.\n\n``workspace_id`` (optional) is stamped into the snapshot metadata so the\nsnapshot read path can enforce tenancy (AAEO-4).",
        "properties": {
          "build_spec": {
            "$ref": "#/components/schemas/ContextBuildSpec"
          },
          "metadata": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata"
          },
          "subject_id": {
            "title": "Subject Id",
            "type": "string"
          },
          "task_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Task Id"
          },
          "workspace_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workspace Id"
          }
        },
        "required": [
          "subject_id",
          "build_spec"
        ],
        "title": "BuildSnapshotRequest",
        "type": "object"
      },
...
      "ContextBuildSpec": {
        "description": "A declarative recipe: which keys to fetch, from where, with what preference.",
        "properties": {
          "keys": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ContextSpecKey"
            },
            "title": "Keys",
            "type": "array"
          },
          "metadata": {
            "additionalProperties": true,
            "default": {},
            "title": "Metadata",
            "type": "object"
          },
          "spec_id": {
            "title": "Spec Id",
            "type": "string"
          }
        },
        "title": "ContextBuildSpec",
        "type": "object"
      },
      "ContextField": {
        "description": "One typed key-value with confidence, freshness, source, and evidence.",
        "properties": {
          "confidence": {
            "default": 1.0,
            "title": "Confidence",
            "type": "number"
          },
          "evidence_refs": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/EvidenceRef"
            },
            "title": "Evidence Refs",
            "type": "array"
          },
          "freshness_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Freshness Seconds"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "metadata": {
            "additionalProperties": true,
            "default": {},
            "title": "Metadata",
            "type": "object"
          },
          "source": {
            "default": "storage",
            "title": "Source",
            "type": "string"
          },
          "value": {
            "title": "Value"
          }
        },
        "required": [
          "key"
        ],
        "title": "ContextField",
        "type": "object"
      },
      "ContextSnapshot": {
        "description": "A frozen, evidenced context result. First-class entity (``kind=\"context_snapshot\"``).\n\nTwo runs with the same snapshot produce comparable answers \u2014 the basis for\nreproducibility, evaluation, and audit.",
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "fields": {
            "additionalProperties": {
              "$ref": "#/components/schemas/ContextField"
            },
            "default": {},
            "title": "Fields",
            "type": "object"
          },
          "metadata": {
            "additionalProperties": true,
            "default": {},
            "title": "Metadata",
            "type": "object"
          },
          "missing_required_keys": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Missing Required Keys",
            "type": "array"
          },
          "snapshot_id": {
            "title": "Snapshot Id",
            "type": "string"
          },
          "subject_id": {
            "title": "Subject Id",
            "type": "string"
          },
          "task_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Task Id"
          }
        },
        "required": [
          "subject_id"
        ],
        "title": "ContextSnapshot",
        "type": "object"
      },
      "ContextSourcePreference": {
        "description": "Where a key's value should be sourced from when building a snapshot.",
        "enum": [
          "storage_first",
          "tool_first",
          "tool_only"
        ],
        "title": "ContextSourcePreference",
        "type": "string"
      },
      "ContextSpecKey": {
        "description": "One entry in a build spec: key, required-ness, source preference, adapter.",
        "properties": {
          "adapter_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Adapter Name"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "metadata": {
            "additionalProperties": true,
            "default": {},
            "title": "Metadata",
            "type": "object"
          },
          "required": {
            "default": false,
            "title": "Required",
            "type": "boolean"
          },
          "source_preference": {
            "$ref": "#/components/schemas/ContextSourcePreference",
            "default": "storage_first"
          }
        },
        "required": [
          "key"
        ],
        "title": "ContextSpecKey",
        "type": "object"
      },
...
      "ErrorResponse": {
        "description": "The consistent error envelope returned for handled failures (AAEO-9).",
        "properties": {
          "code": {
            "default": "error",
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "EvidenceRef": {
        "description": "A pointer to where a context value came from \u2014 a row, tool result, or chunk.",
        "properties": {
          "account_scope": {
            "additionalProperties": true,
            "default": {},
            "title": "Account Scope",
            "type": "object"
          },
          "evidence_id": {
            "title": "Evidence Id",
            "type": "string"
          },
          "metadata": {
            "additionalProperties": true,
            "default": {},
            "title": "Metadata",
            "type": "object"
          },
          "row_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Row Id"
          },
          "source_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Id"
          },
          "source_type": {
            "title": "Source Type",
            "type": "string"
          }
        },
        "required": [
          "source_type"
        ],
        "title": "EvidenceRef",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "x-framework-generated": "validation-error-envelope"
      },
...
      "UpsertFieldsRequest": {
        "description": "Bulk context-field upsert payload for one subject.",
        "properties": {
          "fields": {
            "items": {
              "$ref": "#/components/schemas/ContextField"
            },
            "title": "Fields",
            "type": "array"
          },
          "subject_id": {
            "title": "Subject Id",
            "type": "string"
          },
          "workspace_id": {
            "title": "Workspace Id",
            "type": "string"
          }
        },
        "required": [
          "workspace_id",
          "subject_id",
          "fields"
        ],
        "title": "UpsertFieldsRequest",
        "type": "object"
      },
      "UpsertResult": {
        "description": "The result of a context-field bulk upsert.",
        "properties": {
          "subject_id": {
            "title": "Subject Id",
            "type": "string"
          },
          "upserted": {
            "title": "Upserted",
            "type": "integer"
          }
        },
        "required": [
          "upserted",
          "subject_id"
        ],
        "title": "UpsertResult",
        "type": "object"
      },
...
    }
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/context/fields": {
      "get": {
        "description": "List the stored context fields for a subject (workspace-scoped, AAEO-4).\n\nObject-level (BOLA, WS-bola): a ``subject_scoped`` principal reading ANOTHER data\nsubject's PII fields within its own tenant gets an empty list (the same not-found-shaped\nresponse a foreign subject yields) \u2014 :func:`authorize_object` is a no-op for offline /\n``all_tenants`` / ``tenant_admin`` callers, so the zero-config read is byte-unchanged.",
        "operationId": "list_fields_v1_context_fields_get",
        "parameters": [
          {
            "in": "query",
            "name": "subject_id",
            "required": true,
            "schema": {
              "title": "Subject Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "workspace_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Workspace Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ContextField"
                  },
                  "title": "Response List Fields V1 Context Fields Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Fields",
        "tags": [
          "context"
        ]
      }
    },
    "/v1/context/fields:upsert": {
      "post": {
        "description": "Bulk-upsert context fields for a subject; returns the saved count.\n\nObject-level (BOLA, WS-bola): a ``subject_scoped`` principal may only write under its\nOWN ``subject_id`` (a foreign subject is 403 via :func:`enforce_subject_write`), so a\nsubject-scoped tenant can never overwrite/poison another data subject's PII fields in a\nshared workspace. The gate is a no-op for the offline / ``all_tenants`` / ``tenant_admin``\npath, so the zero-config write is byte-unchanged.",
        "operationId": "upsert_fields_v1_context_fields_upsert_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertFieldsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpsertResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upsert Fields",
        "tags": [
          "context"
        ]
      }
    },
    "/v1/context/snapshots/{snapshot_id}": {
      "get": {
        "description": "Load one context snapshot by id (404 when unknown/out-of-workspace, AAEO-4).\n\nObject-level (BOLA, WS-bola): a ``subject_scoped`` principal resolving a snapshot built\nfor ANOTHER data subject within its own tenant gets a clean 404 \u2014 the workspace scope is\nchecked first, then the snapshot's ``subject_id`` is gated via :func:`authorize_object`.\nA no-op for offline / ``all_tenants`` / ``tenant_admin`` callers (byte-unchanged).",
        "operationId": "get_snapshot_v1_context_snapshots__snapshot_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "snapshot_id",
            "required": true,
            "schema": {
              "title": "Snapshot Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "workspace_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Workspace Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContextSnapshot"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Resource not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Snapshot",
        "tags": [
          "context"
        ]
      }
    },
    "/v1/context/snapshots:build": {
      "post": {
        "description": "Build, persist, and return a context snapshot (stamping workspace scope).\n\nObject-level (BOLA, WS-bola): a ``subject_scoped`` principal may only build a snapshot\nfor its OWN ``subject_id`` (403 otherwise via :func:`enforce_subject_write`) \u2014 a no-op\nfor offline / ``all_tenants`` / ``tenant_admin`` callers.\n\nTenant isolation (red-team reattack-r1): the resolved ``workspace_id`` is also threaded\ninto field RESOLUTION, so a STORAGE-sourced field stamped with a DIFFERENT workspace is\nnever surfaced in the snapshot \u2014 closing the cross-tenant IDOR where two tenants sharing\na free-form ``subject_id`` could read each other's stored ``context_fields`` (the store\nis keyed globally by ``(subject_id, key)``). ``workspace_id`` of ``None`` (offline /\nall-tenants) keeps resolution byte-for-byte unchanged.",
        "operationId": "build_snapshot_v1_context_snapshots_build_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BuildSnapshotRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContextSnapshot"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Build Snapshot",
        "tags": [
          "context"
        ]
      }
    },
...
  }
}